API changes - using fix length field to make integration with "upper" apis easier (how long can a key be? 4 GB is plenty...). Work in progress.
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@885 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/API/CommonObject.h
@@ -68,11 +68,11 @@ typedef struct psoaCommonObject
 
 /** Common function for opening data containers. */
 PHOTON_API_EXPORT
-int psoaCommonObjOpen( psoaCommonObject    * pObject,
+int psoaCommonObjOpen( psoaCommonObject  * pObject,
                       enum psoObjectType   objectType, 
-                      psoaEditMode          editMode,
+                      psoaEditMode         editMode,
                       const char         * objectName,
-                      size_t               nameLengthInBytes );
+                      uint32_t             nameLengthInBytes );
 
 /* --+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+-- */
