+++ src/Nucleus/Queue.c
@@ -103,7 +103,7 @@ void psonQueueFini( psonQueue          * pQueue,
 bool psonQueueGet( psonQueue          * pQueue,
                    unsigned int         flag,
                    psonQueueItem     ** ppIterator,
-                   size_t               bufferLength,
+                   uint32_t             bufferLength,
                    psonSessionContext * pContext )
 {
    psonQueueItem* pQueueItem = NULL;
@@ -242,7 +242,7 @@ bool psonQueueInit( psonQueue           * pQueue,
                     ptrdiff_t             parentOffset,
                     size_t                numberOfBlocks,
                     psonTxStatus        * pTxStatus,
-                    size_t                origNameLength,
+                    uint32_t              origNameLength,
                     char                * origName,
                     ptrdiff_t             hashItemOffset,
                     psoObjectDefinition * pDefinition,
@@ -325,7 +325,7 @@ bool psonQueueInit( psonQueue           * pQueue,
 
 bool psonQueueInsert( psonQueue          * pQueue,
                       const void         * pItem, 
-                      size_t               length ,
+                      uint32_t             length ,
                       enum psonQueueEnum   firstOrLast,
                       psonSessionContext * pContext )
 {
@@ -421,7 +421,7 @@ the_exit:
 
 bool psonQueueInsertNow( psonQueue          * pQueue,
                          const void         * pItem, 
-                         size_t               length ,
+                         uint32_t             length ,
                          enum psonQueueEnum   firstOrLast,
                          psonSessionContext * pContext )
 {
@@ -567,7 +567,7 @@ void psonQueueReleaseNoLock( psonQueue          * pQueue,
 bool psonQueueRemove( psonQueue          * pQueue,
                       psonQueueItem     ** ppQueueItem,
                       enum psonQueueEnum   firstOrLast,
-                      size_t               bufferLength,
+                      uint32_t             bufferLength,
                       psonSessionContext * pContext )
 {
    psoErrors errcode = PSO_OK;
